Move ReceiverWeights tests to a separate file
